requirements.txt
google-cloud-storage>=2.18.0
google-cloud-bigquery>=3.25.0
pandas>=2.2.2
openpyxl>=3.1.5
Unidecode>=1.3.8

